+++ app/pages/index.tsx
@@ -9,6 +9,7 @@ import { Layout } from "../components/Layout";
 import { Modal } from "../components/Modal/Modal";
 import { External_Image } from "../components/External_Image";
 import { MicropostCard } from "../components/Micropost/MicropostCard";
+import { MicropostForm } from "../components/Micropost/MicropostForm";
 //Bootstrap
 import Button from "react-bootstrap/Button";
 import Container from "react-bootstrap/Container";
@@ -30,34 +31,6 @@ const Micropost_Url = process.env.NEXT_PUBLIC_BASE_URL + "microposts";
 
 export default function Home() {
   //State一覧
-  const [errorContent, setErrorContent] = useState<string>("");
-  //投稿画像データを所持するstate
-  const [micropostImage, setMicropostImage] = useState<File>();
-
-  //登校画像データのpreviewを表示
-  const getPreviewMicropostImage = useMemo((): React.ReactElement | void => {
-    if (!micropostImage) {
-      return;
-    }
-    const PreviewUrl = URL.createObjectURL(micropostImage);
-    return <img src={PreviewUrl} className="my-3" style={{ width: "100%" }} />;
-  }, [micropostImage]);
-
-  //写真変更のonChange
-  const handleSetImage = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (!e.target.files) return;
-    const imageFile: File = e.target.files[0];
-    setMicropostImage(imageFile);
-  };
-
-  //input fileをButtonで押す
-  const handleClickInputFile = () => {
-    const target = document.getElementById("image");
-    if (!target) {
-      return;
-    }
-    target.click();
-  };
 
   //useFlashReducerを読み込み
   const { FlashReducer } = useFlashReducer();
@@ -74,62 +47,6 @@ export default function Home() {
 
   //useFeedFetchを読み込み
   const { handleFetching, reloadFetching, isHavingMicroposts } = useFeedFetch();
-
-  //useForm関連メソッド
-  const { register, handleSubmit } = useForm();
-  const onSubmit = (value: { content: string }): void => {
-    const formData = new FormData();
-    formData.append("micropost[content]", value.content);
-    if (micropostImage) {
-      formData.append("micropost[image]", micropostImage);
-    }
-
-    const config = {
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        Authorization: `Bearer ${Auth.getToken()}`,
-      },
-      body: formData,
-    };
-
-    fetch(Micropost_Url, config)
-      .then((response) => {
-        if (!response.ok) {
-          response.json().then((res): any => {
-            if (res.hasOwnProperty("message")) {
-              //authentication関連のエラー処理
-              FlashReducer({ type: "DANGER", message: res.message });
-            }
-          });
-        } else {
-          return response.json();
-        }
-      })
-      .then((data) => {
-        //statusがokayでなければ、dataがundefinedになる
-        if (data == undefined) {
-          return;
-        }
-        // console.log({ data });
-        // Micropostimageを初期化
-        setMicropostImage(undefined);
-        reloadFetching();
-        FlashReducer({ type: "SUCCESS", message: data.message });
-      })
-      .catch((error) => {
-        console.error(error);
-        FlashReducer({ type: "DANGER", message: "Error" });
-      });
-  };
-
-  // const onTestSubmit = (value) => {
-  //   const formData = new FormData()
-  //   formData.append('micropost[content]', value.content)
-  //   formData.append('micropost[image]', micropostImage)
-  //   console.log(formData.get('micropost[content]'))
-  //   console.log(formData.get('micropost[image]'))
-  // }
 
   const FeedScrollList = useMemo(() => {
     return (
@@ -166,7 +83,8 @@ export default function Home() {
           <link rel="icon" href="/favicon.ico" />
         </Head>
 
-        {Auth.isLoggedIn() && user_data && has_user_key() && user_data.user.activated && (
+        {/* {Auth.isLoggedIn() && user_data.user.activated && ( */}
+        {Auth.isLoggedIn() && user_data?.user?.activated && (
           <Container>
             <Row>
               <Col md={5}>
@@ -213,57 +131,7 @@ export default function Home() {
                       </Col>
                     </Row>
                   )}
-                  <Row>
-                    <div className="mt-3 p-3" style={{ width: "100%" }}>
-                      <form onSubmit={handleSubmit(onSubmit)} style={{ maxWidth: "500px" }}>
-                        <textarea
-                          id="content"
-                          // name="content"
-                          placeholder="What's happening to you ?"
-                          style={{
-                            maxWidth: "500px",
-                            height: "100px",
-                            fontSize: "16px",
-                            width: "100%",
-                          }}
-                          className="p-2 "
-                          {...register("content", { required: true })}
-                        />
-                        {errorContent && (
-                          <>
-                            <br />
-                            <p className="text-danger m-0">{errorContent}</p>
-                          </>
-                        )}
-                        <input
-                          className="my-2"
-                          type="file"
-                          accept="image/*"
-                          name="image"
-                          id="image"
-                          style={{ display: "none" }}
-                          onChange={(e: React.ChangeEvent<HTMLInputElement>) => handleSetImage(e)}
-                        />
-                        {getPreviewMicropostImage}
-                        <Button
-                          style={{ maxWidth: "300px", width: "70%" }}
-                          variant="outline-secondary"
-                          onClick={handleClickInputFile}
-                        >
-                          ファイルを選択
-                        </Button>
-
-                        <Button
-                          style={{ maxWidth: "500px", width: "100%" }}
-                          className="mt-3"
-                          variant="outline-primary"
-                          type="submit"
-                        >
-                          Submit
-                        </Button>
-                      </form>
-                    </div>
-                  </Row>
+                  <MicropostForm />
                 </Container>
               </Col>
               <Col md={7}>
